// src/hooks/orval/model/tournamentsRecordStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */

export type TournamentsRecordStatus = typeof TournamentsRecordStatus[keyof typeof TournamentsRecordStatus];


export const TournamentsRecordStatus = {
  draft: 'draft',
  upcoming: 'upcoming',
  live: 'live',
  completed: 'completed',
  archived: 'archived',
} as const;